+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Version 67
 
 * 🔧 On Android 8 and newer, use system settings to configure notification settings.
 * 🔧 On Android 13 and newer, ask to allow notifications.
+* 🔨 Do not open comments screen if required data is not available, yet.
 
 #### 67.0.0 🧪
 *2023-01-27*